Handle Null and Spaces
Sometimes dates are blank (support expiration for rental licenses, or license expiry for socket purchases).  This handles handle the nulls better.  I did so by either matching the support expiration to the license expiration, or setting the expiration to a date far in the future with the new $NoExpiration variable (currently set to Jan 1, 2030).  Sure, its hard coded, but if we are still running this exact script in 2030, I don't think this will be our biggest problem.
Some Nulls are due to old (v9.5 and before) licensing vs new (v10+) licensing.  I used the company name to fill in these nulls.
Additionally, there are some occurrences where the tenant name, and bottleneck may have a space in the name.  I have escaped those spaces.
Finally, just in case there is a comma in the backup job name, we're going to escape that character too.